DATACASS-127 - Minor cleanups in pom.xml.
Changed id of spring.io Artifactory to make sure the project can be prepared using the release shell. Introduce dedicated version for EL as the equality with the CDI version is by accident and might break the build if we update the CDI version going forward. Removed obsolete repositories that are set in the parent POM already.