DATACASS-255 - Consolidate embedded/external cassandra startup.
This commit moves the Cassandra server initialization into one place, the CassandraRule. A test that requires a running Cassandra instance needs to either extend the AbstractEmbeddedCassandraIntegrationTest or add the test rule CassandraRule to its test. A CassandraRule can run initializing and cleanup code that is useful when creating keyspaces or tables.