DATACASS-400 - Use custom conversion and UDT mapping with entity instantiator.
We now apply custom conversion and UDT mapping to the value provider that is used with entity instantiators. This allows entities to be instantiated with mapped UDTs and using custom conversion with properties that are provided using a constructor.

Previously, columns were not converted which caused a parameter mismatch so entities couldn't be instantiated.