DATACASS-568 - Guard BasicCassandraPersistentProperty.findAnnotatedType(…) against non-parameterized types.
We now check whether the inspected property type is a AnnotatedParameterizedType one before attempting to extract type parameters. Map-like and Collection-like properties can be generic typed, untyped or subtype of either one and we restrict annotation scanning to generic typed ones.

Previously, we applied a downcast without checking for the actual type which lead to a ClassCastException.