Allow keyspace customization for Cassandra tables and UDTs.
We now allow setting the keyspace on `@Table` and `@UserDefinedType` to interact with tables and types residing in a specific keyspace.

@Table(keyspace = "some_ks")
class Person {
 …
}

Keyspace values can be either string literals or value expressions (#{…}, ${…}) to use the context or config properties to compute the actual keyspace.

Special honors go to @tomekl007 and @mipo256 for their contributions.

See #1400
See #279

Closes #921