Fix prepared statement caching using observability.
We now no longer create a statement proxy when preparing CQL statements as prepared statements do not use RequestTracker for completion callbacks. Also, ObservationStatement now implements equals and hashCode to report equality for its underlying statement in case the statement was used as cache key.

Previously, we created a proxy without implementing equals and hashCode resulting in re-preparation as the prepared statement cache kept growing because the input statement did not provide means to serve as cache key.

Closes #1601